+++ src/r1cs/mod.rs
@@ -1,5 +1,4 @@
-
-use super::utils::*;
+use crate::utils::*;
 use ark_ff::{Field};
 use thiserror::Error;
